requirements.txt
absl-py==0.7.1
astor==0.7.1
gast==0.2.2
google-pasta==0.1.6
graphviz==0.10.1
grpcio==1.20.1
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
numpy==1.16.3
opencv-python==4.1.0.25
pkg-resources==0.0.0
protobuf==3.7.1
pydotplus==2.0.2
pyparsing==2.4.0
six==1.12.0
tb-nightly==1.14.0a20190301
tensorflow-gpu==2.0.0a0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
Werkzeug==0.15.4
absl-py==0.7.1
astor==0.7.1
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.6
graphviz==0.10.1
grpcio==1.20.1
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1
matplotlib==3.0.3
numpy==1.16.3
opencv-python==4.1.0.25
pandas==0.24.2
pkg-resources==0.0.0
protobuf==3.7.1
pydotplus==2.0.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
tb-nightly==1.14.0a20190301
tensorflow-gpu==2.0.0a0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
Werkzeug==0.15.4
absl-py==0.7.1
astor==0.7.1
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.6
graphviz==0.10.1
grpcio==1.20.1
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1
matplotlib==3.0.3
numpy==1.16.3
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
pkg-resources==0.0.0
protobuf==3.7.1
pydotplus==2.0.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
tb-nightly==1.14.0a20190301
tensorflow-gpu==2.0.0a0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
Werkzeug==0.15.4
absl-py==0.7.1
astor==0.7.1
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.6
graphviz==0.10.1
grpcio==1.20.1
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1
matplotlib==3.0.3
numpy==1.16.3
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
pkg-resources==0.0.0
protobuf==3.7.1
pydotplus==2.0.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
tb-nightly==1.14.0a20190301
tensorflow-gpu==2.0.0a0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
tqdm==4.32.1
Werkzeug==0.15.4
absl-py==0.7.1
astor==0.7.1
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.6
graphviz==0.10.1
grpcio==1.20.1
h5py==2.9.0
joblib==0.13.2
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1
matplotlib==3.0.3
numpy==1.16.3
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
pkg-resources==0.0.0
protobuf==3.7.1
pydotplus==2.0.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
scikit-learn==0.21.1
scipy==1.3.0
six==1.12.0
sklearn==0.0
tb-nightly==1.14.0a20190301
tensorflow-gpu==2.0.0a0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
tqdm==4.32.1
Werkzeug==0.15.4
